Site analysis script with command line arguments
