+++ realistic/load_sose_data.jl
@@ -1,4 +1,3 @@
-using Logging
 using PyCall
 using Conda
